#620: App plugin refactor take 2
